Refactor daemon code, implement daemon iface for greenplum
